+++ src/app/components/ChatInput/index.tsx
@@ -378,6 +378,8 @@ class ChatInput extends React.Component<IProps, IState> {
 
         this.userRepo = UserRepo.getInstance();
 
+        const user = props.peer && (props.peer.getType() === PeerType.PEERUSER && props.peer.getType() === PeerType.PEEREXTERNALUSER) ? this.userRepo.getInstant(props.peer.getId() || '') : null;
+
         this.state = {
             botKeyboard: false,
             disableAuthority: 0x0,
@@ -395,10 +397,10 @@ class ChatInput extends React.Component<IProps, IState> {
             selectable: false,
             selectableDisable: false,
             selectableHasPending: false,
-            showInput: props.peer ? props.peer.getId() !== '2374' : true,
+            showInput: user ? user.id === '2374' || user.deleted : false,
             textareaValue: '',
             uploadPreviewOpen: false,
-            user: props.peer && (props.peer.getType() === PeerType.PEERUSER && props.peer.getType() === PeerType.PEEREXTERNALUSER) ? this.userRepo.getInstant(props.peer.getId() || '') : null,
+            user,
             voiceMode: 'up',
         };
 
@@ -450,7 +452,7 @@ class ChatInput extends React.Component<IProps, IState> {
             this.setState({
                 disableAuthority: 0x0,
                 peer,
-                showInput: peer ? peer.getId() !== '2374' : true,
+                showInput: user ? !(user.id === '2374' || user.deleted) : true,
                 user,
             }, () => {
                 this.checkAuthority()();
@@ -459,6 +461,7 @@ class ChatInput extends React.Component<IProps, IState> {
                 this.userRepo.get(peer.getId() || '').then((res) => {
                     if (res) {
                         this.setState({
+                            showInput: res ? !(res.id === '2374' || res.deleted) : true,
                             user: res,
                         });
                     }
@@ -538,7 +541,7 @@ class ChatInput extends React.Component<IProps, IState> {
             const user = (peer.getType() === PeerType.PEERUSER || peer.getType() === PeerType.PEEREXTERNALUSER) ? this.userRepo.getInstant(peer.getId() || '') : null;
             this.setState({
                 peer,
-                showInput: peer ? peer.getId() !== '2374' : true,
+                showInput: user ? !(user.id === '2374' || user.deleted) : true,
                 user,
             }, () => {
                 this.checkAuthority()();
@@ -547,6 +550,7 @@ class ChatInput extends React.Component<IProps, IState> {
                 this.userRepo.get(peer.getId() || '').then((res) => {
                     if (res) {
                         this.setState({
+                            showInput: res ? !(res.id === '2374' || res.deleted) : true,
                             user: res,
                         });
                     }
@@ -2297,6 +2301,7 @@ class ChatInput extends React.Component<IProps, IState> {
         const {user} = this.state;
         if (user) {
             user.is_bot_started = started;
+            user.dont_update_last_modified = true;
             this.userRepo.importBulk(false, [user]);
         }
     }